moving the JobRunnableWithCancel to the core-framework jobs package
