+++ app/channel/utility/src/main/java/org/phoebus/channelfinder/utility/RemoveTagChannelsJob.java
@@ -4,6 +4,7 @@ import org.phoebus.channelfinder.ChannelFinderClient;
 import org.phoebus.channelfinder.Tag;
 import org.phoebus.framework.jobs.Job;
 import org.phoebus.framework.jobs.JobManager;
+import org.phoebus.framework.jobs.JobRunnableWithCancel;
 
 import java.util.Collection;
 import java.util.function.BiConsumer;
